Add RocksDB dependency and implement blockchain persistence using RocksDB
