+++ crates/core/k2/Cargo.toml
@@ -4,9 +4,7 @@ edition = "2021"
 version.workspace = true
 
 [dependencies]
-
-
-
+rocksdb.workspace = true
 digest.workspace = true
 serde.workspace = true
 hex.workspace = true
@@ -25,10 +23,32 @@ dirs.workspace = true
 tokio.workspace = true
 log.workspace = true    
 env_logger.workspace = true
-consensus-pos.workspace = true
-network.workspace = true
 bcs.workspace = true
 
-move-core-types.workspace = true
 
+move-bytecode-verifier.workspace = true
+move-disassembler.workspace = true
+move-docgen.workspace = true
+move-command-line-common.workspace = true
+move-bytecode-utils.workspace = true
+move-coverage.workspace = true
+move-core-types.workspace = true
+move-ir-types.workspace = true
+move-compiler.workspace = true
+move-stdlib.workspace = true
+move-stdlib-natives.workspace = true
+move-vm-types.workspace = true
+move-vm-runtime.workspace = true
+move-vm-profiler.workspace = true
+move-vm-test-utils.workspace = true
+move-read-write-set-types.workspace = true
+move-binary-format.workspace = true
+move-package.workspace = true
+move-prover.workspace = true
+move-unit-test.workspace = true
+move-errmapgen.workspace = true
+move-bytecode-viewer.workspace = true
+
+consensus-pos.workspace = true
+network.workspace = true
